Extract existing saved-chat history hooks
