+++ research/322-current-app/browser-after.txt
@@ -1,0 +1,41 @@
+- region "Notifications alt+T" [ref=e1]
+  - listitem "Chat renamed successfully" [level=1, ref=e2] focusable [tabindex]
+- link "ChatJS ChatJS" [ref=e3]
+- button [ref=e4]
+- link "New Chat New Chat Cmd+Shift+O" [ref=e5]
+- button "Search chats Cmd+K" [ref=e10]
+- link "Models" [ref=e11]
+- button "New project" [ref=e20]
+- link "Extracted History Verified" [ref=e21]
+- button "More" [expanded=false, ref=e22]
+- button "Dev User Dev User 0.50" [expanded=false, ref=e12]
+- link "Open documentation" [ref=e13]
+- link "Git" [ref=e14]
+- heading "How can I help you today?" [level=1, ref=e6]
+- button "Choose Files" [ref=e15]: No file chosen
+- form "Send a message...GPT-5 miniTools" [ref=e18] clickable [onclick, tabindex]
+  - button "Choose Files" [ref=e23]: No file chosen
+  - textbox [ref=e29]:
+
+  - group "GPT-5 miniTools" [ref=e24] clickable [onclick]
+    - button [expanded=false, ref=e30]
+    - combobox [expanded=false, ref=e31]: openai logo GPT-5 mini
+    - button "Tools" [expanded=false, ref=e32]
+    - button "Submit" [disabled, ref=e33]
+- button "Write" [ref=e25]
+- button "Learn" [ref=e26]
+- button "Code" [ref=e27]
+- button "Life stuff" [ref=e28]
+- button "Open Tanstack query devtools" [ref=e7]
+- button "Open Next.js Dev Tools" [expanded=false, ref=e19]
+- button "icon-close" [ref=e16]
+- button "Clear all events" [ref=e34]
+- button "Primitive.div 48ms" [ref=e35]
+- button "SegmentViewNode 24ms" [ref=e36]
+- button "Primitive.div 40ms" [ref=e37]
+- button "SegmentViewNode 32ms" [ref=e38]
+- button [ref=e40]
+- button "Enable audio alerts" [ref=e39]
+- button "icon-inspect" [ref=e8]
+- button "Notifications" [ref=e9]
+- checkbox "Outline Re-renders" [checked=true, ref=e17]
