+++ research/322-current-app/compilation.json
@@ -1,0 +1,2 @@
+event: message
+data: {"result":{"content":[{"type":"text","text":"{\"issues\":[]}"}]},"jsonrpc":"2.0","id":4}
