+++ research/322-current-app/errors.json
@@ -1,0 +1,2 @@
+event: message
+data: {"result":{"content":[{"type":"text","text":"{\"configErrors\":[],\"sessionErrors\":[]}"}]},"jsonrpc":"2.0","id":3}
