+++ research/322-current-app/expiry-cache-observation.txt
@@ -1,0 +1,36 @@
+- region "Notifications alt+T" [ref=e1]
+- link "ChatJS ChatJS" [ref=e2]
+- button [ref=e3]
+- link "New Chat New Chat Cmd+Shift+O" [ref=e4]
+- button "Search chats Cmd+K" [ref=e7]
+- link "Models" [ref=e8]
+- button "New project" [ref=e19]
+- link "Extracted History Verified" [ref=e20]
+- button "More" [expanded=false, ref=e21]
+- button "Dev User Dev User 0.50" [expanded=false, ref=e9]
+- link "Open documentation" [ref=e10]
+- link "Git" [ref=e11]
+- heading "How can I help you today?" [level=1, ref=e5]
+- button "Choose Files" [ref=e12]: No file chosen
+- form "Send a message...GPT-5 miniTools" [ref=e15] clickable [onclick, tabindex]
+  - button "Choose Files" [ref=e22]: No file chosen
+  - textbox [ref=e28]:
+
+  - group "GPT-5 miniTools" [ref=e23] clickable [onclick]
+    - button [expanded=false, ref=e29]
+    - combobox [expanded=false, ref=e30]: openai logo GPT-5 mini
+    - button "Tools" [expanded=false, ref=e31]
+    - button "Submit" [disabled, ref=e32]
+- button "Write" [ref=e24]
+- button "Learn" [ref=e25]
+- button "Code" [ref=e26]
+- button "Life stuff" [ref=e27]
+- button "Open Tanstack query devtools" [ref=e6]
+- button "Open Next.js Dev Tools" [expanded=false, ref=e16]
+- button "icon-close" [ref=e17]
+- button "Clear all events" [ref=e33]
+- button [ref=e35]
+- button "Enable audio alerts" [ref=e34]
+- button "icon-inspect" [ref=e13]
+- button "Notifications" [ref=e14]
+- checkbox "Outline Re-renders" [checked=true, ref=e18]
