+++ research/322-current-app/unit-tests.log
@@ -1,0 +1,10 @@
+$ vitest run
+The plugin "vite-tsconfig-paths" is detected. Vite now supports tsconfig paths resolution natively via the resolve.tsconfigPaths option. You can remove the plugin and set resolve.tsconfigPaths: true in your Vite config instead.
+
+ RUN  v4.1.4 /private/tmp/chatjs-322-history-current/apps/chat
+
+
+ Test Files  32 passed (32)
+      Tests  149 passed (149)
+   Start at  08:27:59
+   Duration  1.75s (transform 7.68s, setup 0ms, import 12.82s, tests 925ms, environment 8ms)
